Refactor: ponytail pass — drop dead tour surface, fold domain helper
Deletes the uncalled endTour(), the indistinguishable 'live' stage, two
unused seed exports, and the paint-then-override on the popover prev
button; defaultDomainBase moves next to its only caller in
organisation_service.go.
