+++ frontend/src/pages/stacks/onboarding/tour.css
@@ -28,14 +28,17 @@
 
 .driver-popover.stackdome-tour button.driver-popover-next-btn,
 .driver-popover.stackdome-tour button.driver-popover-prev-btn {
-  background-color: var(--brand);
-  color: var(--primary-foreground);
-  border: none;
   border-radius: 9999px;
   padding: 0.3rem 0.85rem;
   font-size: 0.75rem;
   font-weight: 500;
   text-shadow: none;
+}
+
+.driver-popover.stackdome-tour button.driver-popover-next-btn {
+  background-color: var(--brand);
+  color: var(--primary-foreground);
+  border: none;
 }
 
 .driver-popover.stackdome-tour button.driver-popover-prev-btn {
