+++ pkg/services/organisation_service.go
@@ -3,6 +3,9 @@ package services
 import (
 	"context"
 	"fmt"
+	"net"
+	"net/url"
+	"os"
 	"strings"
 
 	"github.com/Stackdome/stackdome/pkg/auth"
@@ -152,6 +155,33 @@ func (s *organisationService) seedPlatformInfra(ctx context.Context, orgID, orgN
 		return sErr
 	}
 	return s.seedOrgRegistry(ctx, orgID, orgName, platformCluster.ID)
+}
+
+// defaultDomainBase resolves the base domain used to seed a self-hosted
+// org's first domain, from APP_BASE_URL in the environment. Hostnames pass
+// through; localhost and IPs are wrapped in nip.io so
+// `<anything>.<ip>.nip.io` resolves back to the same machine. Empty when
+// APP_BASE_URL is unset — no domain is seeded.
+func defaultDomainBase() string {
+	raw := os.Getenv("APP_BASE_URL")
+	if raw == "" {
+		return ""
+	}
+	host := raw
+	if u, err := url.Parse(raw); err == nil && u.Host != "" {
+		host = u.Host
+	}
+	if h, _, err := net.SplitHostPort(host); err == nil {
+		host = h
+	}
+	host = strings.ToLower(host)
+	if host == "localhost" {
+		return "127.0.0.1.nip.io"
+	}
+	if ip := net.ParseIP(host); ip != nil {
+		return host + ".nip.io"
+	}
+	return host
 }
 
 func (s *organisationService) seedOrgDomain(ctx context.Context, orgID, orgSlug, baseDomain string) *errors.ServiceError {
